learnCards refactor:  - replace constants
