+++ src/pages/learnCards/constants.ts
